ZO-3106: Move IPublicationDependencies to cms package from workflow
