+++ core/src/zeit/cms/repository/tests/test_folder.py
@@ -2,6 +2,7 @@ import zope.copypastemove.interfaces
 
 import zeit.cms.testing
 import zeit.cms.interfaces
+import zeit.cms.workflow.interfaces
 import zeit.workflow.interfaces
 import zeit.workflow.testing
 
@@ -21,7 +22,7 @@ class FolderDependenciesTest(zeit.workflow.testing.FunctionalTestCase):
     def test_folder_dependencies(self):
         folder = zeit.cms.interfaces.ICMSContent(
             'http://xml.zeit.de/online/2007')
-        dependencies = list(zeit.workflow.interfaces.IPublicationDependencies(
+        dependencies = list(zeit.cms.workflow.interfaces.IPublicationDependencies(
             folder).get_dependencies())
         self.assertEqual(
             'http://xml.zeit.de/online/2007/01/', dependencies[0].uniqueId)
