+++ core/src/zeit/workflow/interfaces.py
@@ -85,22 +85,6 @@ class IOldCMSStatus(zope.interface.Interface):
         required=False)
 
 
-class IPublicationDependencies(zope.interface.Interface):
-    """Adapter to find the publication dependencies of an object."""
-
-    def get_dependencies():
-        """Return a sequence of all dependent objects.
-
-    The sequence contains all objects which need to be published along with the
-    adapted object. Dependent containers will be published recursively.
-    """
-
-    retract_dependencies = zope.interface.Attribute("""\
-        If True, retract dependent objects along with the adapted object.
-        Usually we cannot know whether an object is used by someone else and
-        thus can't retract it, but in some cases this decision can be made.""")
-
-
 class IPublisherData(zope.interface.Interface):
     """
     Adapter to return json data for content for specific third party systems
