+++ core/src/zeit/workflow/publish.py
@@ -245,7 +245,7 @@ class PublishRetractTask:
                 break
 
             # Dive into dependent objects
-            deps = zeit.workflow.interfaces.IPublicationDependencies(new_obj)
+            deps = zeit.cms.workflow.interfaces.IPublicationDependencies(new_obj)
             if self.mode == MODE_PUBLISH:
                 stack.extend(deps.get_dependencies())
             elif self.mode == MODE_RETRACT:
